refactor(server): split giant process file into smaller files
